Additional core functions

+++ runtime/eval_test.go
@@ -9,12 +9,11 @@ import (
 )
 
 func TestEval(t *testing.T) {
-	// TODO: always returning int64 is not great
-	assert.Equal(t, int64(3), evalStr("(+ 1 2)"))
+	assert.Equal(t, 3, evalStr("(+ 1 2)"))
 }
 
 func TestRecursiveEval(t *testing.T) {
-	assert.Equal(t, int64(10), evalStr("(+ (+ 1 2) (+ 3 4))"))
+	assert.Equal(t, 10, evalStr("(+ (+ 1 2) (+ 3 4))"))
 }
 
 func evalStr(s string) interface{} {
